[FIX] move functions to auth blueprint
```login```
url has changed to /auth/login

```logout```
url has changed to /auth/logout

```get_logged_in_username```
now you should import this function from AwesomeTitleServer.auth.utils
